06 - 複製/iam_system/requirements.txt
Flask==2.0.1
PyJWT==2.4.0
